fix: terminal animation component
